// src/generated/models/KnowYourMarginReq.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type KnowYourMarginReq = {
    /**
     * User specific identification generated by Dhan
     */
    dhanClientId?: string;
    /**
     * | **Enums**    | **Exchange** | **Segment**       |
     * |--------------|--------------|-------------------|
     * | NSE_EQ       | NSE          | Equity Cash       |
     * | NSE_FNO      | NSE          | Futures & Options |
     * | BSE_EQ       | BSE          | Equity Cash       |
     * | BSE_FNO      | BSE          | Futures & Options |
     * | MCX_COMM     | MCX          | Commodity         |
     *
     */
    exchangeSegment: KnowYourMarginReq.exchangeSegment;
    /**
     * Signifies the type of transaction whether it's BUY or SELL
     */
    transactionType: KnowYourMarginReq.transactionType;
    /**
     * Number of shares for the order
     */
    quantity?: number;
    /**
     * | **Enum Values** | **Description**                                |
     * |-----------------|------------------------------------------------|
     * | CNC             | Cash & Carry for equity deliveries             |
     * | INTRADAY        | Intraday for Equity, Futures & Options         |
     * | MARGIN          | Carry Forward in Futures & Options             |
     * | CO              | Cover Order; entry and stop loss               |
     * | BO              | Bracket Order; entry, stop loss & target price |
     * **CO & BO product types will be valid only for Intraday**
     */
    productType?: KnowYourMarginReq.productType;
    /**
     * Exchange standard identification for each scrip
     */
    securityId?: string;
    /**
     * Price at which the order is requested to execute
     */
    price?: number;
    /**
     * Price at which the order is requested to trigger
     */
    triggerPrice?: number;
};
export namespace KnowYourMarginReq {
    /**
     * | **Enums**    | **Exchange** | **Segment**       |
     * |--------------|--------------|-------------------|
     * | NSE_EQ       | NSE          | Equity Cash       |
     * | NSE_FNO      | NSE          | Futures & Options |
     * | BSE_EQ       | BSE          | Equity Cash       |
     * | BSE_FNO      | BSE          | Futures & Options |
     * | MCX_COMM     | MCX          | Commodity         |
     *
     */
    export enum exchangeSegment {
        NSE_EQ = 'NSE_EQ',
        NSE_FNO = 'NSE_FNO',
        BSE_EQ = 'BSE_EQ',
        BSE_FNO = 'BSE_FNO',
        MCX_COMM = 'MCX_COMM',
    }
    /**
     * Signifies the type of transaction whether it's BUY or SELL
     */
    export enum transactionType {
        BUY = 'BUY',
        SELL = 'SELL',
    }
    /**
     * | **Enum Values** | **Description**                                |
     * |-----------------|------------------------------------------------|
     * | CNC             | Cash & Carry for equity deliveries             |
     * | INTRADAY        | Intraday for Equity, Futures & Options         |
     * | MARGIN          | Carry Forward in Futures & Options             |
     * | CO              | Cover Order; entry and stop loss               |
     * | BO              | Bracket Order; entry, stop loss & target price |
     * **CO & BO product types will be valid only for Intraday**
     */
    export enum productType {
        CNC = 'CNC',
        INTRADAY = 'INTRADAY',
        MARGIN = 'MARGIN',
        MTF = 'MTF',
        CO = 'CO',
        BO = 'BO',
    }
}